relocataing ingredeints to single component;
