+++ src/components/BurgerConstructor/BurgerConstructorStyles.module.scss
@@ -25,21 +25,6 @@
       overflow-y: scroll;
       min-height: 30px;
     }
-
-    .constructor_order_item {
-      padding-right: 10px;
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-
-      svg {
-        margin-right: 10px;
-      }
-
-      & > svg:nth-child(1) {
-        cursor: pointer;
-      }
-    }
   }
 
   .price_info {
