+++ src/components/BurgerConstructor/CurrentIngredients/CurrentIngredientsStyles.module.scss
@@ -1,0 +1,14 @@
+.constructor_order_item {
+  padding-right: 10px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+
+  svg {
+    margin-right: 10px;
+  }
+
+  & > svg:nth-child(1) {
+    cursor: pointer;
+  }
+}
